refactor(sveltevietnam.dev+kit): extract SOCIAL_LINKS and EMAILS constants to kit
