--- sites/sveltevietnam.dev/src/data/emails.ts
@@ -1,11 +1,0 @@
-export const EMAILS = {
-	ADMIN: 'admin@sveltevietnam.dev',
-	JOBS: 'jobs@sveltevietnam.dev',
-	BLOG: 'blog@sveltevietnam.dev',
-	CONTACT: 'contact@sveltevietnam.dev',
-	COC: 'coc@sveltevietnam.dev',
-	EVENTS: 'events@sveltevietnam.dev',
-	IMPACT: 'impact@sveltevietnam.dev',
-	SPONSOR: 'sponsor@sveltevietnam.dev',
-} as const;
-
