+++ sites/sveltevietnam.dev/src/data/links.ts
@@ -1,12 +1,3 @@
-export const SOCIAL_LINKS = {
-	GITHUB: 'https://github.com/sveltevietnam',
-	DISCORD: 'https://discord.sveltevietnam.dev',
-	BLUESKY: 'https://bsky.app/profile/sveltevietnam.dev',
-	FACEBOOK: 'https://facebook.com/sveltevietnam',
-	OPEN_COLLECTIVE: 'https://opencollective.com/sveltevietnam',
-	YOUTUBE: 'https://www.youtube.com/@sveltevietnam',
-} as const;
-
 export const GITHUB_LINKS = {
 	PROJECT: 'https://github.com/orgs/sveltevietnam/projects/1',
 	PROJECT_REFERENCES:
