+++ sites/sveltevietnam.dev/src/routes/(site)/[lang=lang]/jobs/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { T } from '@sveltevietnam/i18n/runtime';
 	import { TBA, Breadcrumbs } from '@sveltevietnam/kit/components';
+	import { EMAILS } from '@sveltevietnam/kit/constants';
 	import { Contexts } from '@sveltevietnam/kit/contexts';
 
 	import { page } from '$app/state';
-	import { EMAILS } from '$data/emails';
 	import * as m from '$data/locales/generated/messages';
 	import * as p from '$data/routes/generated';
 	import { GradientBackground } from '$lib/components/gradient-background';
